server_py/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1

